+++ kitty/options/parse.py
@@ -1201,6 +1201,9 @@ class Parser:
 
     choices_for_pointer_shape_when_grabbed = choices_for_default_pointer_shape
 
+    def progress_bar(self, val: str, ans: dict[str, typing.Any]) -> None:
+        ans['progress_bar'] = to_bool(val)
+
     def remember_window_position(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['remember_window_position'] = to_bool(val)