+++ kitty/options/types.py
@@ -415,6 +415,7 @@ option_names = (
     'placement_strategy',
     'pointer_shape_when_dragging',
     'pointer_shape_when_grabbed',
+    'progress_bar',
     'remember_window_position',
     'remember_window_size',
     'remote_control_password',
@@ -622,6 +623,7 @@ class Options:
     placement_strategy: choices_for_placement_strategy = 'center'
     pointer_shape_when_dragging: tuple[str, str] = ('beam', 'crosshair')
     pointer_shape_when_grabbed: choices_for_pointer_shape_when_grabbed = 'arrow'
+    progress_bar: bool = True
     remember_window_position: bool = False
     remember_window_size: bool = True
     repaint_delay: int = 10