+++ kitty/screen.h
@@ -169,6 +169,8 @@ typedef struct {
         unsigned int val;
     } prompt_settings;
     char display_window_char;
+    uint8_t progress_state;   // 0=unset 1=set 2=error 3=indeterminate 4=paused
+    uint8_t progress_percent; // 0-100
     struct {
         char ch;
         uint8_t *canvas;